[FEATURE] fs.readFile recursively searches modules
If fs.readFile is given a path including "contract_modules/{module}" it
will check to see if that module is included in the contract's manifest.
If the module is included, it will recursively search that module's
included files and submodules until the file is found.
Modified test contract to demonstrate this feature.

*Note: need to handle ".", "..", and leading "/" in paths better