[FEATURE] Automatically compile contract from dir
compile-contract.js script compiles a contract from a directory
It also compiles submodules and automatically generates the manifests.
Manifests are populated with the name (based on the directory name),
the main file (either contract.js, index.js, or main.js), the files,
the module dependencies, and the APIs used.
All of the dependency files are written to the contract_filesystem
using the hashes of the data as the filenames.

Updated test_contract and run-test-contract.js script to use compile-contract